==> personalize/1. user-interactions-clean.ipynb <==
#### Persionalize 구성 흐름
user interaction 최소 조건
- At minimum 1000 interactions records from users interacting with items in your catalog.
- At minimum 25 unique user IDs with at least 2 interactions for each.

1. 데이터 탐색
2. 데이터 정제
3. S3 Bucket 생성, Policy 설정,  정제 데이터 업로드


import pandas as pd
import datetime
pd.set_option('display.width', 1000)

data_df = pd.read_csv('./data/data.csv')
data_df.head()

print(data_df.shape)

# instraction datasetd에 필수 항목 추출
insteractions_df = data_df[['user_id', 'product_id','event_time', 'event_type']]
# items_df = data_df[['product_id', 'price', 'brand', 'cat_0', 'cat_1', 'cat-2']]
print(insteractions_df.shape)
insteractions_df.isnull().sum()

# dataframe 내에 colname 이 multi_cnt 이상 df 반환
def filter_df(df, colname='user_id',  multi_cnt=2):
    # colname user_id 인 경우  아래 조건을 맞족한는 Series의 index 가 user_id
    is_multi = df[colname].value_counts() > 1
    # Series boolean index를 이용
    multi_index = is_multi[is_multi].index
    filtered = df[df[colname].isin(multi_index)]
    return filtered, multi_index

#  최소 거래건수: 2개 이상
filtered_df, user_ids_atleast_2event = filter_df(insteractions_df)
print(filtered_df.shape)


filtered_df.info()

- event_time:  object --> Unix timestamp 타입 변환 필요

### evnet_time type conversion: object to Unix timestamp

user_interactions_df = filtered_df.copy()

# 유효하지 않는  format은  NaT으로 변환 및 제거: object -> datetime
user_interactions_df['event_time'] = pd.to_datetime(user_interactions_df['event_time'], format="%Y-%m-%d %H:%M:%S UTC", errors='coerce')
user_interactions_df = user_interactions_df[pd.notnull(user_interactions_df['event_time'])]


# event_time의 datetime을 Unix timestamp 타입으로
user_interactions_df['event_time'] = user_interactions_df.event_time.apply(lambda x : (x-datetime.datetime(1970,1,1)).total_seconds())
user_interactions_df.info()

# convert event_time from float to int
user_interactions_df['event_time'] = user_interactions_df['event_time'].astype(int)

user_interactions_df.info()


# rename columne:  Personalize를 위한 필수 항목: USER_ID, ITEM_ID, TIMESTAMP
user_interactions_df.rename(columns={"user_id": "USER_ID", "product_id": "ITEM_ID", "event_time": "TIMESTAMP", "event_type": "EVENT_TYPE" }, inplace=True)


user_interactions_df.isnull().sum()

# 정제된 데이터를 파일로 저장
user_interactions_df.to_csv("./data/cleaned_user_interactions.csv", index=False)
