==> personalize/3. solution.ipynb <==
#### Solution 생성 및 Versioning
솔루선 생성 및 하이퍼 파라미터 조정

import boto3
from botocore.exceptions import ClientError
import json
import time


personalize = boto3.client("personalize")

#### 솔루션 생성

# recipe (알고리즘) 선택
recipe_arn = "arn:aws:personalize:::recipe/aws-user-personalization"

# 이미 dataset group이 생성되어 있을 때, 조회
# dataset_group_name = "osungmart-dataset-groupmwomk"
dataset_group_name = "osungmart-dataset-group20230204"

list_dataset_group_resp = personalize.list_dataset_groups(
  maxResults=10
)

for dataset_grp in list_dataset_group_resp['datasetGroups']:
  if dataset_grp['name'] == dataset_group_name:
    dataset_group_arn = dataset_grp['datasetGroupArn']
    print("dataset_group_arn: {}".format(dataset_grp['datasetGroupArn']) )

# create solution
solution_name = "osugmart-user-personalize-solution-HPO"
try: 
  create_solution_resp = personalize.create_solution(
    name = solution_name,
    datasetGroupArn = dataset_group_arn,
    recipeArn = recipe_arn,
    performHPO=True
  )
  solution_arn = create_solution_resp['solutionArn']
  print(json.dumps(create_solution_resp, indent=2))
except ClientError as e:
  print(e)

solution_arn = "arn:aws:personalize:ap-northeast-2:491818659652:solution/osugmart-user-personalize-solution-HPO"

# Create Solution Version: training and retraining
try:
  create_solution_version_resp = personalize.create_solution_version(
    solutionArn = solution_arn
  )
  solution_version_arn = create_solution_version_resp['solutionVersionArn']
  print(json.dumps(create_solution_version_resp, indent=2))
except ClientError as e:
  print(e)


**Solution Versioning은 많은 시간이 소요되므로, 모니터링은 Console에서 확인하는 것이 좋다**

status = None
max_time = time.time() + 3*60*60 # 3 hours
while time.time() < max_time:
    describe_solution_version_response = personalize.describe_solution_version(
        solutionVersionArn = solution_version_arn
    )
    status = describe_solution_version_response["solutionVersion"]["status"]
    print("SolutionVersion: {}".format(status))
    
    if status == "ACTIVE" or status == "CREATE FAILED":
        break
        
    time.sleep(60)

#### 성능 평가
AWS User Personalization metrics

personalize_runtime = boto3.client('personalize-runtime')
personalize_events = boto3.client("personalize-events")

user insteraction 2건 이상, 10 만건은 metric이 매우 좋지 않았음

# 2023. 02. 07. 2nd. user interaction: 2건, 100만건
solution_version_arn = "arn:aws:personalize:ap-northeast-2:491818659652:solution/osugmart-user-personalize-solution-HPO/afda062c"

get_solution_metrics_response = personalize.get_solution_metrics(
    solutionVersionArn = solution_version_arn
)
print(json.dumps(get_solution_metrics_response, indent=2))

# 2023. 02. 08. 3th. user interactin: 5, 200만건
solution_version_arn = "arn:aws:personalize:ap-northeast-2:491818659652:solution/osugmart-user-personalize-solution-HPO/90a54ee1"

get_solution_metrics_response = personalize.get_solution_metrics(
    solutionVersionArn = solution_version_arn
)
print(json.dumps(get_solution_metrics_response, indent=2))
