==> personalize/4. Campaign.ipynb <==
import boto3
from botocore.exceptions import ClientError
import json
import time
import pandas as pd
import numpy as np
import uuid

personalize = boto3.client("personalize")
personalize_runtime = boto3.client('personalize-runtime')
personalize_events = boto3.client("personalize-events")

#### Campaign 생성

Campaign은 배포된 Solution Version입니다. 실질적을 추천서비스를 제공하는 endpoint입니다.
가격은 처리율 즉 TPS에 따라 결정되며, 일반적으로 테스트시에는 1로 설정합니다.

배포에는 약 10여분 정도가 소요됩니다

# 배포할 Solution Version
solution_version_arn = "arn:aws:personalize:ap-northeast-2:491818659652:solution/osugmart-user-personalize-solution-HPO/90a54ee1"

suffix = "20230209"
campaign_name = "osungmart-campaign-" + suffix

try:
    create_campaign_resp = personalize.create_campaign(
        name = campaign_name,
        solutionVersionArn = solution_version_arn,
        minProvisionedTPS = 1
    )
    
    campaign_arn = create_campaign_resp['campaignArn']
    print(json.dumps(create_campaign_resp, indent=2))
except ClientError as e:
    if e.response["Error"]["Code"] == "ResourceAlreadyExistsException":
        print(f"dataset group: {dataset_group_name} arleady exist")
    else:
        raise


# Campaign 생성 모니터링
status = None
max_time = time.time() + 3*60*60 # 3 hours
while time.time() < max_time:
    describe_campaign_resp = personalize.describe_campaign(
        campaignArn = campaign_arn
    )
    status = describe_campaign_resp["campaign"]["status"]
    print("Campaign: {}".format(status))
    
    if status == "ACTIVE" or status == "CREATE FAILED":
        break
        
    time.sleep(60)

#### 추천
cleand data sample에서 user id를 추출해서 사용자에 따른 상품 추천을 테스트한다.

cleaned_insteractions = pd.read_csv('data/cleaned_user_interactions.csv')
cleaned_insteractions.head()

cleaned_insteractions['EVENT_TYPE'].unique()

# 3 user id sampling
users = cleaned_insteractions.sample(n=3)['USER_ID'].to_list()
users

#campaign_arn = ""

# get Recommendation
get_recommendation_resp = personalize_runtime.get_recommendations(
    campaignArn = campaign_arn,
    userId = str(users[0])
)
item_list = get_recommendation_resp['itemList']
print(item_list)

#### Real time events Simulation

event_tracker_name = "Osung-Event-Tracker"
dataset_group_arn = "arn:aws:personalize:ap-northeast-2:491818659652:dataset-group/osungmart-dataset-group20230204"

# event tracker 생성
resp = personalize.create_event_tracker(
    name = event_tracker_name,
    datasetGroupArn = dataset_group_arn
)
print("eventTrackerArn: ", resp['eventTrackerArn'])
print("trackingId", resp['trackingId'])
TRACKING_ID = resp['trackingId']
event_tracker_arn = resp['eventTrackerArn']

##### method for simulations

compaign_arn = "arn:aws:personalize:ap-northeast-2:491818659652:campaign/osungmart-campaign-20230209"

session_dict = {}
def send_event(USER_ID, ITEM_ID, EVENT_TYPE ='purchase'):
    """
    Simulates a click as an envent
    to send an event to Amazon Personalize's Event Tracker
    """
    # Configure Session
    try:
        session_ID = session_dict[str(USER_ID)]
    except:
        session_dict[str(USER_ID)] = str(uuid.uuid1())
        session_ID = session_dict[str(USER_ID)]
    
    # Configure Properties
    event = {
        "itemId": str(ITEM_ID)
    }
    
    event_json = json.dumps(event)
    # Make a Call
    personalize_events.put_events(
        trackingId = TRACKING_ID,
        userId = str(USER_ID),
        sessionId = session_ID,
        eventList = [{
            "sentAt": int(time.time()),
            "eventType": EVENT_TYPE,
            "properties": event_json
        }]
    )
            
def get_new_recommendation_uers_real_time(user_id, item_id_list):
    # Send event
    for item_id in item_id_list:
        send_event(USER_ID=user_id, ITEM_ID=item_id)
        time.sleep(5)
    
    # get recommendation
    get_recommendation_resp = personalize_runtime.get_recommendations(
        campaignArn = compaign_arn,
        userId = str(user_id),
    )
    
    item_list = get_recommendation_resp['itemList']
    recommendation_list = []
    for item in item_list:
        recommendation_list.append(item['itemId'])
    
    return recommendation_list
    

# cart event 가 많은 사용자 추출
event_group = cleaned_insteractions.groupby(by=['USER_ID', 'EVENT_TYPE']).count().reset_index()
# 10번 이상 cart event가 발생한 user
user_cart_10_more = event_group.query('(EVENT_TYPE =="cart") and ( TIMESTAMP >= 10)')['USER_ID']
user_cart_10_more= user_cart_10_more.tolist()
len(user_cart_10_more)

# 기존 데이터 셋에서 user가 cart넣은 상품을 10샘플링해서 구매(purchase) event로 발생
user_id = user_cart_10_more[1]
print('user id: ', user_id)
item_id_list = cleaned_insteractions.query('USER_ID == @user_id')['ITEM_ID'].tolist()
item_id_list

recommendataion_list = get_new_recommendation_uers_real_time(user_id, item_id_list[:5])

recommendataion_list
